pyproject.toml
[project]
name = "nexus-madrid"
version = "0.1.0"
description = "Workstation analítica de inteligencia inmobiliaria para Madrid"
requires-python = ">=3.13,<3.14"

dependencies = [
  "PySide6==6.10.2",
  "SQLAlchemy==2.0.48",
  "alembic==1.18.4",
  "polars==1.38.1",
  "pydantic==2.12.5",
  "scikit-learn==1.8.0",
  "RapidFuzz==3.14.3",
  "numpy>=2.1,<3.0",
  "python-dateutil>=2.9,<3.0",
  "orjson>=3.10,<4.0",
  "loguru>=0.7,<1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.6,<1.0",
  "black>=24.0,<26.0",
  "mypy>=1.10,<2.0"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]